requirements.txt
streamlit
langchain
openai
python-dotenv
groq